Refactor to ClaimingSummaary component
